requirements.txt
Flask==1.1.1
Flask_RESTful==0.3.7
faiss-cpu==1.6.1
boto3==1.11.0
pytest==5.3.2
python-dotenv==0.12.0
